Reorganize asset paths code
