+++ dandiapi/api/asset_paths.py
@@ -1,12 +1,48 @@
 from __future__ import annotations
 
+import os
+
 from django.db import transaction
-from django.db.models import F, QuerySet
+from django.db.models import Count, F, QuerySet
 
 from dandiapi.api.models import Asset, AssetPath, AssetPathRelation, Version
 
-from .query import get_path_children, get_root_paths
-from .utils import extract_paths
+
+def extract_paths(path: str) -> list[str]:
+    nodepaths: list[str] = path.split('/')
+    for i in range(len(nodepaths))[1:]:
+        nodepaths[i] = os.path.join(nodepaths[i - 1], nodepaths[i])
+
+    return nodepaths
+
+
+def get_root_paths(version: Version) -> QuerySet[AssetPath]:
+    """Return all root paths for a version."""
+    # Use prefetch_related here instead of select_related,
+    # as otherwise the resulting join is very large
+    qs = AssetPath.objects.prefetch_related(
+        'asset',
+        'asset__blob',
+        'asset__embargoed_blob',
+        'asset__zarr',
+    )
+    return qs.filter(version=version).alias(num_parents=Count('parent_links')).filter(num_parents=1)
+
+
+def get_path_children(path: AssetPath) -> QuerySet[AssetPath]:
+    """Get all direct children from an existing path."""
+    qs = AssetPath.objects.select_related(
+        'asset',
+        'asset__blob',
+        'asset__embargoed_blob',
+        'asset__zarr',
+    )
+    path_ids = (
+        AssetPathRelation.objects.filter(parent=path, depth=1)
+        .values_list('child', flat=True)
+        .distinct()
+    )
+    return qs.filter(id__in=path_ids)
 
 
 def search_asset_paths(query: str, version: Version) -> QuerySet[AssetPath] | None:
